Fully extract all entries from groups
