+++ src/lib/template/seaweed/entries/EntryProps.ts
@@ -1,0 +1,13 @@
+export enum DefaultHeader {
+	Games = "Games",
+	Projects = "Projects"
+}
+
+export interface EntryProps {
+	email: string;
+}
+
+export interface GroupedEntry {
+	header: string;
+	entryList: EntryProps[];
+}
