+++ src/routes/(seaweed)/portfolio/+page.svelte
@@ -1,9 +1,6 @@
 <script>
 	import SeaweedTemplate from "$pkg/template/seaweed/SeaweedTemplate.svelte";
-
-	export let metaWebsite = "https://www.crouton.com/test/portfolio-base-layout";
-
 </script>
 
 <!--todo: maybe add check that this is always false? -->
-<SeaweedTemplate letChaos={true}></SeaweedTemplate>
+<SeaweedTemplate letChaos={false}></SeaweedTemplate>
